Extract CGIEscapeTest from CGIUtilTest
